split out axis type defaults and calculate before anything else
